Opdracht12

+++ SoftwareDevelopment/Module-02/robotarm-python-2021-main/RobotArmBasics1.py
@@ -1,19 +1,27 @@
 from RobotArm import RobotArm
 
-robotArm = RobotArm('exercise 5')
+robotArm = RobotArm('exercise 12')
 
 # Jouw python instructies zet je vanaf hier:
-for i in range(7):
-    robotArm.moveRight()
 
-for i in range(8):
+Rijen = 9
+
+while True:
+    if Rijen == 0:
+        break
     robotArm.grab()
-    robotArm.moveRight()
-    robotArm.drop()
-    robotArm.moveLeft()
-    if i == 8:
-        print()
-    elif i < 8:
-        robotArm.moveLeft()
+    Kleur = robotArm.scan()
+    if Kleur != "red":
+        robotArm.drop()
+        robotArm.moveRight()
+        Rijen -= 1
+    else:
+        for i in range(Rijen):
+            robotArm.moveRight()
+        robotArm.drop()
+        for i in range(Rijen-1):
+            robotArm.moveLeft()
+        Rijen -= 1
+
 # Na jouw code wachten tot het sluiten van de window:
 robotArm.wait()
